Clanker build

+++ src/components/damage.h
@@ -6,7 +6,7 @@
 #include "../ecs/entity.h"
 
 struct Damage{
-    uint32_t damage_counters;
+    size_t damage_counters;
 };
 
 bool deal_damage(Entity source, Entity target, size_t amount);
